Move upgradeCurrencyCodes, makeCurrencyCodeTable -> tokenIdTools.ts
