+++ src/actions/ScanActions.tsx
@@ -6,7 +6,7 @@ import URL from 'url-parse'
 import { selectWalletForExchange } from '../actions/CryptoExchangeActions'
 import { ButtonsModal } from '../components/modals/ButtonsModal'
 import { ConfirmContinueModal } from '../components/modals/ConfirmContinueModal'
-import { upgradeCurrencyCodes, WalletListModal, WalletListResult } from '../components/modals/WalletListModal'
+import { WalletListModal, WalletListResult } from '../components/modals/WalletListModal'
 import { Airship, showError, showWarning } from '../components/services/AirshipInstance'
 import { getSpecialCurrencyInfo } from '../constants/WalletAndCurrencyConstants'
 import { lstrings } from '../locales/strings'
@@ -17,7 +17,8 @@ import { NavigationBase } from '../types/routerTypes'
 import { getTokenId } from '../util/CurrencyInfoHelpers'
 import { parseDeepLink } from '../util/DeepLinkParser'
 import { logActivity } from '../util/logger'
-import { getPluginIdFromChainCode, makeCurrencyCodeTable, toListString, zeroString } from '../util/utils'
+import { makeCurrencyCodeTable, upgradeCurrencyCodes } from '../util/tokenIdTools'
+import { getPluginIdFromChainCode, toListString, zeroString } from '../util/utils'
 import { cleanQueryFlags, openBrowserUri } from '../util/WebUtils'
 
 /**
